AddTokenForm and debounce search
