+++ src/components/pages-component/home-page/index.ts
@@ -1,0 +1,4 @@
+export * from './add-token-form';
+export * from './modal-center';
+export * from './modal-left-side';
+export * from './other';
