+++ src/components/pages-component/home-page/modal-center.tsx
@@ -1,0 +1,26 @@
+import type { ReactNode } from 'react';
+
+export const ModalCenter = ({
+  open,
+  children,
+  onClose,
+}: {
+  open: boolean;
+  children: ReactNode;
+  onClose: () => void;
+}) => {
+  if (!open) return null;
+  return (
+    <>
+      <div className="fixed inset-0 bg-black/50" onClick={onClose} />
+      <div className="fixed inset-x-1/2 inset-y-1/4 z-10">
+        <div className="flex w-96 -translate-x-1/2 -translate-y-1/2 flex-col rounded-lg bg-white p-4">
+          <button type="button" className="mb-4 text-left" onClick={onClose}>
+            X
+          </button>
+          {children}
+        </div>
+      </div>
+    </>
+  );
+};
